requirements.txt
colorama
python-nmap
nvdlib
rich
